requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap-base-template==0.0.2
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
iso-639==0.4.5
google-cloud-translate==1.6.0